Resolve code duplication for classification/regression summarizers
